use management network for contrail control plane
Split contrail/vip in two specific network roles: contrail/mgmt on wich all management trafic
will be bound, and contrail/data for overlay trafic on the vrouter.

These two networks are still bound to private network by default, so that it doesn't change
anything in typical deployment, but it enables users to split management and dataplane
trafic using network templates.

contrail_private_vip should be renamed to contrail_mgmt_vip in various templates, but it is
preferable to it in a subsequent (sed) patchset, it was ommited for now for the sake of clarity